Update code to fit new database structre
The new database has a few changes.

There is no more a table called VIETA there are 2
seperate tables now called MIESTAS and MIKRORAJONAS.
The address that was fitted into the VIETA table is
now in the RENGINYS table for clarity.

Code added is for the most part to adapt to the new
database.