Make microcity_dropdown work
Made it so that in the event_create page the microcity_dropdown
works.

After 4 hours it finally works.

If you start the page from the /IT_darbas/public/ page
then all the sripts will be working from there, so will
the php code so if you are doing a relative path be
aware of that.