Fix messages not found not showing in messages
Before I used to go staright to the moodule rather
than through the controller and because of that
even when there were no messages it would still have
an array with someting from the query.
That is removed in the controller.

Removed error log from Event module because it's no
longer needed.