Add updateTabeRowPopularity in Subscription
Now when creating a Subscription and choosing what kind
of event you want, with what kind of people and where
that key word will get a plus one in it's usage.

The 4 tables RENGINIO_TIPAS, MIESTAS, MIKRORAJONAS
and SOCIALINES_GRUPES have a value called kartu_panaudotas.
This will be used to show the user how popular a certain
key word is.