Add red cirle to messages if not read
Made it so that if the user has unread messages then
the messages tab will have a red circle.

Also made it so that the message itself would have
a red circle also. If the message is pressed then
the red circle will be removed from the message and
if there are no more unread messages then the red
cirlce in the tab will also disappear.

The ajax and mark_as_read is confusing...