Fix wrong submission when filtering
Before POST was used when filtering events. That
made it so that when going to the event that you
searched for and then back to the filtering page
there would be a need to resubmit the form.

With GET the data is saved in the page paramaters
and so works even if you go between pages.