Bug fix: Subscription and messages interaction
Before when a user had a Subscription that did not have any
event types or social groups then it would be ignored even
if other paramaters matched when an event was created.

Now if a user creates a Subscription that does not have
event types or social groups he will still receive messages
about events that match the other conditions.